Add make_pr.sh [skip ci]
To make it easier to create a PR when the further analysis
reveals that the packages are OK to be added.

The checker should say which packages are available in the
source package.
All one has to do is run:

    ./make_pr.sh apt-package-whitelist 1234

If the file named `package` exists in the same directory,
its content is read to pick up which packages to add.